# docs/api/openapi.json
{
  "components": {
    "schemas": {
      "AgentRunIn": {
        "additionalProperties": false,
        "properties": {
          "ticket_text": {
            "anyOf": [
              {
                "maxLength": 5000,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Ticket Text"
          }
        },
        "title": "AgentRunIn",
        "type": "object"
      },
      "AuthorizationIn": {
        "additionalProperties": false,
        "properties": {
          "authorized_amount": {
            "anyOf": [
              {
                "exclusiveMinimum": 0.0,
                "type": "number"
              },
              {
                "pattern": "^(?!^[-+.]*$)[+-]?0*(?:\\d{0,18}|(?=[\\d.]{1,19}0*$)\\d*\\.\\d*0*$)",
                "type": "string"
              }
            ],
            "title": "Authorized Amount"
          },
          "currency": {
            "default": "INR",
            "pattern": "^[A-Za-z]{3}$",
            "title": "Currency",
            "type": "string"
          },
          "customer_id": {
            "pattern": "^[A-Za-z0-9][A-Za-z0-9_.:-]{0,63}$",
            "title": "Customer Id",
            "type": "string"
          },
          "metadata": {
            "additionalProperties": true,
            "title": "Metadata",
            "type": "object"
          },
          "operation": {
            "$ref": "#/components/schemas/Operation",
            "default": "REFUND"
          },
          "order_id": {
            "pattern": "^[A-Za-z0-9][A-Za-z0-9_.:-]{0,63}$",
            "title": "Order Id",
            "type": "string"
          },
          "ticket_text": {
            "anyOf": [
              {
                "maxLength": 5000,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Ticket Text"
          }
        },
        "required": [
          "customer_id",
          "order_id",
          "authorized_amount"
        ],
        "title": "AuthorizationIn",
        "type": "object"
      },
      "CancelIn": {
        "additionalProperties": false,
        "properties": {
          "note": {
            "default": "",
            "maxLength": 2000,
            "title": "Note",
            "type": "string"
          }
        },
        "title": "CancelIn",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "Operation": {
        "enum": [
          "REFUND",
          "PAYMENT_AUTHORIZATION"
        ],
        "title": "Operation",
        "type": "string"
      },
...
      "ProposalIn": {
        "additionalProperties": false,
        "properties": {
          "agent_id": {
            "anyOf": [
              {
                "maxLength": 64,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Agent Id"
          },
          "amount": {
            "anyOf": [
              {
                "exclusiveMinimum": 0.0,
                "type": "number"
              },
              {
                "pattern": "^(?!^[-+.]*$)[+-]?0*(?:\\d{0,18}|(?=[\\d.]{1,19}0*$)\\d*\\.\\d*0*$)",
                "type": "string"
              }
            ],
            "title": "Amount"
          },
          "currency": {
            "default": "INR",
            "pattern": "^[A-Za-z]{3}$",
            "title": "Currency",
            "type": "string"
          },
          "customer_id": {
            "pattern": "^[A-Za-z0-9][A-Za-z0-9_.:-]{0,63}$",
            "title": "Customer Id",
            "type": "string"
          },
          "operation": {
            "$ref": "#/components/schemas/Operation"
          },
          "order_id": {
            "pattern": "^[A-Za-z0-9][A-Za-z0-9_.:-]{0,63}$",
            "title": "Order Id",
            "type": "string"
          },
          "rationale": {
            "default": "",
            "maxLength": 2000,
            "title": "Rationale",
            "type": "string"
          },
          "request_id": {
            "anyOf": [
              {
                "maxLength": 64,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Request Id"
          }
        },
        "required": [
          "operation",
          "customer_id",
          "order_id",
          "amount"
        ],
        "title": "ProposalIn",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      }
    }
  },
  "info": {
    "description": "Payment execution, recovery and reconciliation between applications or AI agents and a payment provider. Contract: docs/API.md. Simulated provider only.",
    "title": "IntentGuard Gateway",
    "version": "3.0.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/authorizations": {
      "post": {
        "operationId": "create_authorization_api_authorizations_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AuthorizationIn"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "title": "Response Create Authorization Api Authorizations Post"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Operator authorization: creates the intent",
        "tags": [
          "intents"
        ]
      }
    },
...
    "/api/intents": {
      "get": {
        "operationId": "list_intents_api_intents_get",
        "parameters": [
          {
            "in": "query",
            "name": "state",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "State"
            }
          },
          {
            "in": "query",
            "name": "customer_id",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Customer Id"
            }
          },
          {
            "in": "query",
            "name": "order_id",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Order Id"
            }
          },
          {
            "in": "query",
            "name": "operation",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Operation"
            }
          },
          {
            "in": "query",
            "name": "created_from",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Created From"
            }
          },
          {
            "in": "query",
            "name": "created_to",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Created To"
            }
          },
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "integer"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Limit"
            }
          },
          {
            "in": "query",
            "name": "cursor",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Cursor"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response List Intents Api Intents Get",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "List intents (newest first, cursor-paginated)",
        "tags": [
          "intents"
        ]
      }
    },
    "/api/intents/{intent_id}": {
      "get": {
        "operationId": "get_intent_api_intents__intent_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "intent_id",
            "required": true,
            "schema": {
              "title": "Intent Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response Get Intent Api Intents  Intent Id  Get",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Intent with attempts, effects and the open review case",
        "tags": [
          "intents"
        ]
      }
    },
    "/api/intents/{intent_id}/agent": {
      "post": {
        "operationId": "run_agent_api_intents__intent_id__agent_post",
        "parameters": [
          {
            "in": "path",
            "name": "intent_id",
            "required": true,
            "schema": {
              "title": "Intent Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "anyOf": [
                  {
                    "$ref": "#/components/schemas/AgentRunIn"
                  },
                  {
                    "type": "null"
                  }
                ],
                "title": "Body"
              }
            }
          }
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "title": "Response Run Agent Api Intents  Intent Id  Agent Post"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Extract a proposal from the ticket and submit it",
        "tags": [
          "intents"
        ]
      }
    },
    "/api/intents/{intent_id}/cancel": {
      "post": {
        "operationId": "cancel_api_intents__intent_id__cancel_post",
        "parameters": [
          {
            "in": "path",
            "name": "intent_id",
            "required": true,
            "schema": {
              "title": "Intent Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "anyOf": [
                  {
                    "$ref": "#/components/schemas/CancelIn"
                  },
                  {
                    "type": "null"
                  }
                ],
                "title": "Body"
              }
            }
          }
        },
        "responses": {
          "202": {
            "content": {
              "application/json": {
                "schema": {
                  "title": "Response Cancel Api Intents  Intent Id  Cancel Post"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Cancel (pending effects are verified)",
        "tags": [
          "intents"
        ]
      }
    },
    "/api/intents/{intent_id}/history": {
      "get": {
        "operationId": "history_api_intents__intent_id__history_get",
        "parameters": [
          {
            "in": "path",
            "name": "intent_id",
            "required": true,
            "schema": {
              "title": "Intent Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response History Api Intents  Intent Id  History Get",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Raw rows: proposals, decisions, attempts, effects, ...",
        "tags": [
          "intents"
        ]
      }
    },
    "/api/intents/{intent_id}/proposals": {
      "post": {
        "operationId": "submit_proposal_api_intents__intent_id__proposals_post",
        "parameters": [
          {
            "in": "path",
            "name": "intent_id",
            "required": true,
            "schema": {
              "title": "Intent Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ProposalIn"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "title": "Response Submit Proposal Api Intents  Intent Id  Proposals Post"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Submit a structured proposal (agent boundary)",
        "tags": [
          "intents"
        ]
      }
    },
    "/api/intents/{intent_id}/timeline": {
      "get": {
        "operationId": "timeline_api_intents__intent_id__timeline_get",
        "parameters": [
          {
            "in": "path",
            "name": "intent_id",
            "required": true,
            "schema": {
              "title": "Intent Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response Timeline Api Intents  Intent Id  Timeline Get",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Ordered, human-readable events",
        "tags": [
          "intents"
        ]
      }
    },
...
  }
}